# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: MindAtlas API
  description: AI-powered mapping and organization web application API
  version: 0.1.0
  contact:
    name: MindAtlas Team
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from POST /api/auth/login.
        Tokens expire after 15 minutes; use /api/auth/refresh to renew.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        API key created via POST /api/keys.
        Used for programmatic/webhook access.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid email or password
      required: [error]
...
paths:
...
  /api/admin/users:
    get:
      tags: [admin]
      summary: List users (admin)
      description: Paginated list of user accounts. Requires admin role + MFA.
      operationId: adminListUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled, locked]
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        email:
                          type: string
                        role:
                          type: string
                        status:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/users/{id}:
    get:
      tags: [admin]
      summary: Get user details (admin)
      operationId: adminGetUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details (no Card content)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  status:
                    type: string
                  subscription_tier:
                    type: string
                  created_at:
                    type: string
                    format: date-time
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/users/{id}/disable:
    post:
      tags: [admin]
      summary: Disable a user account
      operationId: adminDisableUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: User disabled
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/users/{id}/delete:
    post:
      tags: [admin]
      summary: Delete a user account
      operationId: adminDeleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: User deleted
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/users/{id}/unlock:
    post:
      tags: [admin]
      summary: Unlock a locked user account
      operationId: adminUnlockUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User unlocked
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/plans:
    get:
      tags: [admin]
      summary: List subscription plans
      operationId: adminListPlans
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Plans list
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        price_cents:
                          type: integer
                        active:
                          type: boolean
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [admin]
      summary: Create a subscription plan
      operationId: adminCreatePlan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price_cents:
                  type: integer
                limits:
                  type: object
              required: [name, price_cents]
      responses:
        '201':
          description: Plan created
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/plans/{id}/entitlements:
    get:
      tags: [admin]
      summary: Get feature entitlements for a plan
      operationId: adminGetEntitlements
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entitlements list
          content:
            application/json:
              schema:
                type: object
                properties:
                  entitlements:
                    type: array
                    items:
                      type: object
                      properties:
                        feature_key:
                          type: string
                        enabled:
                          type: boolean
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [admin]
      summary: Update feature entitlements for a plan
      operationId: adminUpdateEntitlements
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                features:
                  type: array
                  items:
                    type: object
                    properties:
                      feature_key:
                        type: string
                      enabled:
                        type: boolean
      responses:
        '200':
          description: Entitlements updated
        '400':
          description: Invalid feature key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/features:
    get:
      tags: [admin]
      summary: List all registered features
      operationId: adminListFeatures
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Feature registry
          content:
            application/json:
              schema:
                type: object
                properties:
                  features:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        name:
                          type: string
                        category:
                          type: string
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/audit:
    get:
      tags: [admin]
      summary: Get admin audit log
      operationId: adminAuditLog
      security:
        - BearerAuth: []
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: admin_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        admin_id:
                          type: string
                        action:
                          type: string
                        target_id:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/moderate/{userId}:
    post:
      tags: [admin]
      summary: Moderate a user
      description: Flag or disable a user account for policy violations
      operationId: adminModerate
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [flag, disable]
                reason:
                  type: string
              required: [action]
      responses:
        '200':
          description: Moderation action applied
        '400':
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'